refactor(agents): dismantle AgentGenerator God Class into engine modules
